+++ {{cookiecutter.github_repository}}/settings/production.py
@@ -90,6 +90,13 @@ if ENABLE_MEDIA_UPLOAD_TO_S3:
 # Static Assests
 # ------------------------
 STATICFILES_STORAGE = 'whitenoise.storage.CompressedManifestStaticFilesStorage'
+{%- if cookiecutter.add_sass_with_django_compressor.lower() == 'y' %}
+
+# Compress static files offline
+# http://django-compressor.readthedocs.org/en/latest/settings/#django.conf.settings.COMPRESS_OFFLINE
+COMPRESS_STORAGE = 'compressor.storage.GzipCompressorFileStorage'
+COMPRESS_OFFLINE = True
+{%- endif %}
 
 # EMAIL
 # ------------------------------------------------------------------------------